windows fixes
fucking windows REEEEEEEEEEEEEEEEEEEEEEEE
fopen() silently converts line-endings on windows, and since I did
not explicitly say I was wiring binary data, it solently bloated my
sav size.
Literally fixed by adding a single char: "w" -> "wb"

Also cleaned up some things that were causing compiler warnings.
Nothing mission-critical, but its nice to keep things clean.

+++ src/nes/cpu/cpu.cc
@@ -403,14 +403,14 @@ uint CPU::step() {
               } break;
     case ROL: { if (opcode.addrm == Instructions::AddrM::acc) {
                   bool old_bit_0 = nth_bit(this->reg.a, 7);
-                  this->reg.a = (this->reg.a << 1) | this->reg.p.c;
+                  this->reg.a = (this->reg.a << 1) | u8(this->reg.p.c);
                   this->reg.p.c = old_bit_0;
                   set_zn(this->reg.a);
                 } else {
                   u8 val = this->mem[addr];
                   this->mem[addr] = val; // dummy-write
                   bool old_bit_0 = nth_bit(val, 7);
-                  val = (val << 1) | this->reg.p.c;
+                  val = (val << 1) | u8(this->reg.p.c);
                   this->reg.p.c = old_bit_0;
                   set_zn(val);
                   this->mem[addr] = val;
